+++ templates.go
@@ -32,6 +32,8 @@ func baseFuncMap() template.FuncMap {
 			}
 			return m
 		},
+		"add": func(a, b int) int { return a + b },
+		"sub": func(a, b int) int { return a - b },
 	}
 }